+++ scripts/extension-tools/teax.py
@@ -90,6 +90,7 @@ CONTENT_SHEETS = {
     "Cards",
     "CardTags",
     "Relations",
+    "Surveys",
 }